# docs/openapi/platform.yaml
openapi: 3.0.3
info:
  title: DeliveryHub API
  version: v1
  description: See docs/api.md for the conventions (errors, pagination, authentication).
paths:
...
  /api/v1/businesses:
    get:
      operationId: businesses_list
      summary: All businesses with today's numbers
      tags:
      - businesses
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Businesses'
          description: ''
    post:
      operationId: businesses_create
      summary: 'Open a business: its schema, addresses and the owner''s admin account'
      tags:
      - businesses
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BusinessCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BusinessCreateRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessCreated'
          description: ''
...
components:
  schemas:
    Bot:
      type: object
      properties:
        username:
          type: string
        alive:
          type: boolean
          description: The bot service polls it right now
        created_via:
          $ref: '#/components/schemas/CreatedViaEnum'
      required:
      - alive
      - created_via
      - username
...
    Bots:
      type: object
      properties:
        client:
          allOf:
          - $ref: '#/components/schemas/Bot'
          nullable: true
        admin:
          allOf:
          - $ref: '#/components/schemas/Bot'
          nullable: true
      required:
      - admin
      - client
    BusinessCard:
      type: object
      description: 'Needs `request` in the context (addresses follow the style of
        the request: public or *.localhost).'
      properties:
        slug:
          type: string
        name:
          type: string
        status:
          $ref: '#/components/schemas/StatusEnum'
        logo:
          type: string
          nullable: true
        brand_color:
          type: string
        tagline:
          type: string
        created_at:
          type: string
          format: date-time
        links:
          $ref: '#/components/schemas/Links'
        stats:
          $ref: '#/components/schemas/Stats'
        bots:
          $ref: '#/components/schemas/Bots'
      required:
      - bots
      - brand_color
      - created_at
      - links
      - logo
      - name
      - slug
      - stats
      - status
      - tagline
    BusinessCreateRequest:
      type: object
      description: 'Editable profile; `logo: null` (or empty) removes the logo.'
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        tagline:
          type: string
          maxLength: 200
        support_phone:
          type: string
          maxLength: 30
        delivery_time:
          type: string
          maxLength: 20
        min_order:
          type: integer
          maximum: 100000000
          minimum: 0
        brand_color:
          type: string
          maxLength: 7
        logo:
          type: string
          format: binary
          nullable: true
        slug:
          type: string
          minLength: 1
          description: '3–30 characters: a-z, 0-9, "-"'
          maxLength: 40
        owner_name:
          type: string
          minLength: 1
          maxLength: 120
        owner_phone:
          type: string
          minLength: 1
          maxLength: 30
        owner_password:
          description: Empty → generated
          oneOf:
          - type: string
            maxLength: 128
            minLength: 8
          - type: string
            maxLength: 0
      required:
      - name
      - owner_name
      - owner_phone
      - slug
    BusinessCreated:
      type: object
      properties:
        business:
          $ref: '#/components/schemas/BusinessDetail'
        credentials:
          $ref: '#/components/schemas/Credentials'
      required:
      - business
      - credentials
    BusinessDetail:
      type: object
      description: 'Needs `request` in the context (addresses follow the style of
        the request: public or *.localhost).'
      properties:
        slug:
          type: string
        name:
          type: string
        status:
          $ref: '#/components/schemas/StatusEnum'
        logo:
          type: string
          nullable: true
        brand_color:
          type: string
        tagline:
          type: string
        created_at:
          type: string
          format: date-time
        links:
          $ref: '#/components/schemas/Links'
        stats:
          $ref: '#/components/schemas/Stats'
        bots:
          $ref: '#/components/schemas/Bots'
        support_phone:
          type: string
        delivery_time:
          type: string
        min_order:
          type: integer
        owner:
          allOf:
          - $ref: '#/components/schemas/Owner'
          nullable: true
        platform_bot:
          allOf:
          - $ref: '#/components/schemas/PlatformBot'
          nullable: true
        missing_roles:
          type: array
          items:
            $ref: '#/components/schemas/MissingRolesEnum'
      required:
      - bots
      - brand_color
      - created_at
      - delivery_time
      - links
      - logo
      - min_order
      - missing_roles
      - name
      - owner
      - platform_bot
      - slug
      - stats
      - status
      - support_phone
      - tagline
    BusinessTotals:
      type: object
      properties:
        businesses:
          type: integer
        active:
          type: integer
        orders_today:
          type: integer
        revenue_today:
          type: integer
      required:
      - active
      - businesses
      - orders_today
      - revenue_today
    Businesses:
      type: object
      properties:
        domain:
          type: string
          description: 'PLATFORM_DOMAIN: businesses live at <slug>.<domain>'
        totals:
          $ref: '#/components/schemas/BusinessTotals'
        results:
          type: array
          items:
            $ref: '#/components/schemas/BusinessCard'
      required:
      - domain
      - results
      - totals
...
    CreatedViaEnum:
      enum:
      - managed
      - token
      type: string
      description: |-
        * `managed` - managed
        * `token` - token
    Credentials:
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
      required:
      - password
      - phone
...
    Links:
      type: object
      properties:
        shop:
          type: string
          format: uri
        admin:
          type: string
          format: uri
      required:
      - admin
      - shop
...
    MissingRolesEnum:
      enum:
      - client
      - admin
      type: string
      description: |-
        * `client` - Customers
        * `admin` - Staff
    Owner:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
      required:
      - name
      - phone
...
    PlatformBot:
      type: object
      properties:
        username:
          type: string
      required:
      - username
...
    Stats:
      type: object
      properties:
        orders_today:
          type: integer
        revenue_today:
          type: integer
        orders_total:
          type: integer
        customers:
          type: integer
      required:
      - customers
      - orders_today
      - orders_total
      - revenue_today
    StatusEnum:
      enum:
      - active
      - suspended
      type: string
      description: |-
        * `active` - active
        * `suspended` - suspended
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid